- Update: Some Generic love in WorldMgr/Region/Zone
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/eden@3050 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/gameutils/IDoor.cs
@@ -40,7 +40,7 @@ namespace DOL.GS
 		ushort ZoneID { get; }
 		eRealm Realm {get;}
 		int DoorID	{get;}
-		int ObjectID	{get;}
+		ushort ObjectID	{get;}
 		eDoorState State {get; set;}
 		void Open();
 		void Close();
